refactor: migrate sign in google logic to a separate function
